Node generation WIP

+++ shady_generator/src/error.rs
@@ -17,7 +17,7 @@ pub enum ShadyError {
     #[error("Could not find output property with uuid `{0}`")]
     MissingOutputProperty(String),
     #[error("Node processing reached depth {0}, check your nodes for potential loops")]
-    MaxDepthReached(u8),
+    MaxDepthReached(usize),
     #[error("Detected a loop for node {0}")]
     NodeLoopDetected(String),
 }
